Revert "refactor components"
This reverts commit bbf7a2653edf9c64f2297a3a2d23e92fc5abd197.
